Remove all type annotations from skeletons
With zsh extended globbing:

    sed -i 's/::\w*//g' **/(^example).jl~**/runtests.jl

+++ exercises/pascals-triangle/pascals-triangle.jl
@@ -1,3 +1,3 @@
-function triangle(n::Int)
+function triangle(n)
 
 end
